fix: saved images toggle
