+++ src/pages/Saved/Saved.js
@@ -6,9 +6,24 @@ function Saved(props) {
   const [photos, setPhotos] = useLocalStorage('savedImages', []);
   const [data, setData] = useState([]);
 
+  const handleSave = (content) => {
+    const imageExists = photos.find((ph) => ph.id === content.id);
+    if (imageExists) {
+      const photosArr = photos.filter((ph) => ph.id !== content.id);
+      setPhotos(photosArr);
+    }
+    if (!imageExists) {
+      const photosArr = [...photos, content];
+      setPhotos(photosArr);
+    }
+  };
   return (
     <Wrapper>
-      <PhotoList photos={photos} setPhotos={setPhotos} />
+      <PhotoList
+        photos={photos}
+        setPhotos={setPhotos}
+        handleSave={handleSave}
+      />
     </Wrapper>
   );
 }
